Rename math to mathutils (issue #23)
Now the imports are not broken
