+++ qstack/__init__.py
@@ -7,5 +7,5 @@ from qstack import fields
 from qstack import basis_opt
 from qstack import spahm
 from qstack import regression
-from qstack import math
+from qstack import mathutils
 from qstack import orcaio
