+++ qstack/mathutils/fps.py
