+++ qstack/mathutils/xyz_integrals_sym.py
