+++ qstack/regression/hyperparameters.py
@@ -6,7 +6,7 @@ import scipy
 from sklearn.model_selection import train_test_split, KFold
 from qstack.regression.kernel_utils import get_kernel, defaults, ParseKwargs
 from qstack.tools import correct_num_threads
-from qstack.math.fps import do_fps
+from qstack.mathutils.fps import do_fps
 
 def hyperparameters(X, y,
            sigma=defaults.sigmaarr, eta=defaults.etaarr, gkernel=defaults.gkernel, gdict=defaults.gdict,
