+++ tests/test_reorder.py
@@ -4,7 +4,7 @@ import os
 import numpy as np
 from qstack import compound, tools
 from qstack.fields.decomposition import decompose
-from qstack.math.matrix import from_tril
+from qstack.mathutils.matrix import from_tril
 
 
 def test_reorder_pyscf_gpr():
